Move account filter test to SVM
