--- svm/tests/mod.rs
@@ -1,1 +1,0 @@
-mod rent_state;
